Refactors code so main is separate from textql
