+++ textql.go
@@ -1,9 +1,8 @@
-package main
+package textql
 
 import (
 	"database/sql"
 	"encoding/csv"
-	"flag"
 	"fmt"
 	"io/ioutil"
 	"regexp"
@@ -15,7 +14,6 @@ import (
 	"io"
 	"log"
 	"os"
-	"os/exec"
 	"os/user"
 	"path/filepath"
 	"strings"
@@ -23,23 +21,9 @@ import (
 	"unicode/utf8"
 )
 
-func main() {
-	// Parse command line opts
-	commands := flag.String("sql", "", "SQL Command(s) to run on the data")
-	source_text := flag.String("source", "stdin", "Source file to load, or defaults to stdin")
-	delimiter := flag.String("dlm", ",", "Delimiter between fields -dlm=tab for tab, -dlm=0x## to specify a character code in hex")
-	lazyQuotes := flag.Bool("lazy-quotes", false, "Enable LazyQuotes in the csv parser")
-	header := flag.Bool("header", false, "Treat file as having the first row as a header row")
-	outputHeader := flag.Bool("output-header", false, "Display column names in output")
-	tableName := flag.String("table-name", "tbl", "Override the default table name (tbl)")
-	save_to := flag.String("save-to", "", "If set, sqlite3 db is left on disk at this path")
-	console := flag.Bool("console", false, "After all commands are run, open sqlite3 console with this data")
-	verbose := flag.Bool("verbose", false, "Enable verbose logging")
-	flag.Parse()
-
-	if *console && (*source_text == "stdin") {
-		log.Fatalln("Can not open console with pipe input, read a file instead")
-	}
+func Load(source_text *string, delimiter *string, lazyQuotes *bool,
+	header *bool, tableName *string, save_to *string, console *bool,
+	verbose *bool) (*sql.DB, *string, rune) {
 
 	separator := determineSeparator(delimiter)
 
@@ -124,10 +108,15 @@ func main() {
 		fmt.Fprintf(os.Stderr, "Data loaded in: %v\n", t1.Sub(t0))
 	}
 
+	return db, openPath, separator
+}
+
+func Execute(db *sql.DB, commands *string, separator rune, verbose *bool, outputHeader *bool) {
+
 	// Determine what sql to execute
 	sqls_to_execute := strings.Split(*commands, ";")
 
-	t0 = time.Now()
+	t0 := time.Now()
 
 	// Execute given SQL
 	for _, sql_cmd := range sqls_to_execute {
@@ -140,41 +129,10 @@ func main() {
 		}
 	}
 
-	t1 = time.Now()
+	t1 := time.Now()
 
 	if *verbose {
 		fmt.Fprintf(os.Stderr, "Queries run in: %v\n", t1.Sub(t0))
-	}
-
-	// Open console
-	if *console {
-		db.Close()
-		args := []string{*openPath}
-		if *outputHeader {
-			args = append(args, "-header")
-		}
-		cmd := exec.Command("sqlite3", args...)
-
-		cmd.Stdin = os.Stdin
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		cmd_err := cmd.Run()
-		if cmd.Process != nil {
-			cmd.Process.Release()
-		}
-
-		if len(*save_to) == 0 {
-			os.RemoveAll(filepath.Dir(*openPath))
-		}
-
-		if cmd_err != nil {
-			log.Fatalln(cmd_err)
-		}
-	} else if len(*save_to) == 0 {
-		db.Close()
-		os.Remove(*openPath)
-	} else {
-		db.Close()
 	}
 }
